Update README.md
In Communication Research in the Big Data Era: The Application of Machine Learning Models and AI Techniques in Communication Research, Xiaoqun Zhang argues that acquiring knowledge of machine learning (ML) and artificial intelligence (AI) tools is increasingly imperative for the trajectory of communication research in the era of big data. Rather than simply being a matter of keeping pace with technological advances, Zhang posits that these tools are strategically imperative for navigating the complexities of the digital media landscape and big data analysis, and they provide powerful methodologies empowering researchers to uncover nuanced insights and trends within the vast expanse of digital information. Although this can be a daunting notion for researchers without a formal background in mathematics or computer science, this book highlights the substantial rewards of investing time and effort into the endeavor—mastery of ML and AI not only facilitates more sophisticated big data analyses but also fosters interdisciplinary collaborations, enhancing the richness and depth of research outcomes. This book will serve as a foundational resource for communication scholars by providing essential knowledge and techniques to effectively leverage ML and AI at the intersection of communication research and data science.